Leave runtime package as an optional (simplify get_platform_flash_runtime), package extraction in own module
